+++ spec/directions/move_solutions_to_answer_key/calculus_strategy_spec.rb
@@ -103,10 +103,10 @@ RSpec.describe Kitchen::Directions::MoveSolutionsToAnswerKey::V1 do
       <<~HTML
         <div class="os-eob os-solutions-container" data-type="composite-chapter" data-uuid-key=".solutions">
           <h1 data-type="document-title" id="composite-chapter-1">
-            <span class="os-text">#{I18n.t(:eoc_answer_key_title)}</span>
+            <span class="os-text">#{I18n.t(:answer_key_title)}</span>
           </h1>
           <div data-type="metadata" style="display: none;">
-            <h1 data-type="document-title" itemprop="name">#{I18n.t(:eoc_answer_key_title)}</h1>
+            <h1 data-type="document-title" itemprop="name">#{I18n.t(:answer_key_title)}</h1>
             #{metadata(title: 'Answer Key', id_suffix: '_copy_1')}
           </div>
         </div>