+++ CHANGELOG.md
@@ -16,6 +16,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Switched to using a library to sort strings in a language-specific way (patch)
 * Remove summary attribute from `BakeNumberedTable` (major)
 
+* Add `BakeChapterSolutions` which bakes the free response solutions at the eoc (minor)
+* Changed locale `eoc_answer_key_title` to `answer_key_title` as it was only used in eob (major)
+
 ## [4.1.1] - 2021-05-24
 
 * Adds low level Nokogiri caching, disabled by default (patch)